Make setting configs more user-friendly
Previously there were tons of flags that the user must set, and
these are scattered throughout the config.py file. This commit
organizes the necessary settings by category and moves the
important ones to the top.

It also replaces all "SKIP_*" flags with the positive form,
which makes it more straightforward for the user to reason about.